servicios Rest descargar pdf

+++ proyectoEPS/proyectoEPS-ejb/src/main/java/gt/edu/usac/cunoc/ingenieria/eps/project/facade/ProjectFacade.java
@@ -46,7 +46,7 @@ public class ProjectFacade implements ProjectFacadeLocal {
 
     @Override
     public Project getProject(Integer projectId) {
-        return projectRepository.getProjects(projectId.intValue(), null, null).get(0);
+        return projectRepository.getProjects(projectId, null, null).get(0);
     }
 
     @Override
